+++ include/boost/beast/core/buffered_read_stream.hpp
@@ -111,8 +111,7 @@ public:
         typename std::remove_reference<Stream>::type;
 
     /// The type of the lowest layer.
-    using lowest_layer_type =
-        typename get_lowest_layer<next_layer_type>::type;
+    using lowest_layer_type = get_lowest_layer<next_layer_type>;
 
     /** Move constructor.