+++ test/beast/core/type_traits.cpp
@@ -71,8 +71,7 @@ struct F3
     using next_layer_type =
         typename std::remove_reference<F>::type;
 
-    using lowest_layer_type = typename
-        get_lowest_layer<next_layer_type>::type;
+    using lowest_layer_type = get_lowest_layer<next_layer_type>;
 };
 
 template<class F>
@@ -81,18 +80,18 @@ struct F4
     using next_layer_type =
         typename std::remove_reference<F>::type;
 
-    using lowest_layer_type = typename
-        get_lowest_layer<next_layer_type>::type;
+    using lowest_layer_type =
+        get_lowest_layer<next_layer_type>;
 };
 
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F1>::type, F1>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F2>::type, F2>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F3<F1>>::type, F1>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F3<F2>>::type, F2>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F1>>::type, F1>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F2>>::type, F2>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F3<F1>>>::type, F1>::value);
-BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F3<F2>>>::type, F2>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F1>, F1>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F2>, F2>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F3<F1>>, F1>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F3<F2>>, F2>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F1>>, F1>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F2>>, F2>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F3<F1>>>, F1>::value);
+BOOST_STATIC_ASSERT(std::is_same<get_lowest_layer<F4<F3<F2>>>, F2>::value);
 
 //
 // min_all, max_all